backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
python-dotenv==1.0.1
supabase==2.10.0
sentence-transformers==3.3.1
openai==1.57.2
numpy==1.26.4
pydantic==2.10.3
PyPDF2==3.0.1
python-docx==1.1.2
